wip: #60 removed react-table usage from dashboard
