+++ apps/issues/components/ProjectList.tsx
@@ -1,13 +1,5 @@
-'use client';
-
 import { InformationCircleIcon } from '@heroicons/react/outline';
 import { Prisma } from '@prisma/client';
-import {
-  createColumnHelper,
-  flexRender,
-  getCoreRowModel,
-  useReactTable,
-} from '@tanstack/react-table';
 import Image from 'next/image';
 import Link from 'next/link';
 
@@ -21,109 +13,27 @@ const ProfileImage = ({ src, user }: { src?: string; user: string }) => {
   );
 };
 
-export function ProjectList({
-  isLoading,
-  projects,
-}: {
-  isLoading: boolean;
-  projects: ProjectType[] | null | undefined;
-}) {
+export function ProjectList({ projects }: { projects: ProjectType[] }) {
   const dateOptions: Intl.DateTimeFormatOptions = {
     day: 'numeric',
     month: 'long',
     year: 'numeric',
   };
 
-  const columnHelper = createColumnHelper<ProjectType>();
-  const columns = [
-    columnHelper.accessor('name', {
-      header: () => 'Name',
-      cell: (info) => <div className="pl-1">{info.getValue()}</div>,
-    }),
-    columnHelper.accessor('startDate', {
-      header: () => 'Start Date',
-      cell: (info) => {
-        const date = info.getValue();
-
-        return (
-          <div className="text-center">
-            {typeof date === 'undefined' || date === null
-              ? ''
-              : date.toLocaleString('en-US', dateOptions)}
-          </div>
-        );
-      },
-    }),
-    columnHelper.accessor('endDate', {
-      header: () => 'End Date',
-      cell: (info) => {
-        const date = info.getValue();
-
-        return (
-          <div className="text-center">
-            {typeof date === 'undefined' || date === null
-              ? ''
-              : date.toLocaleString('en-US', dateOptions)}
-          </div>
-        );
-      },
-    }),
-    columnHelper.accessor('team', {
-      header: () => 'Team',
-      cell: (info) => (
-        <div className="team-avatars">
-          {info.getValue().map(({ user }) => (
-            <ProfileImage key={user.id} user={user.name ?? user.id} />
-          ))}
-        </div>
-      ),
-    }),
-    columnHelper.display({
-      id: 'ticketCount',
-      header: () => 'Ticket Count',
-      cell: (info) => (
-        <div className="text-center">{info.row.original.tickets.length}</div>
-      ),
-    }),
-    columnHelper.display({
-      id: 'details',
-      cell: (info) => (
-        <Link href={`/projects/${info.row.original.id}`}>
-          <InformationCircleIcon className="h-5 mx-auto w-5 hover:cursor-pointer" />
-        </Link>
-      ),
-    }),
-  ];
-
-  const table = useReactTable({
-    columns,
-    data: projects ?? [],
-    getCoreRowModel: getCoreRowModel(),
-  });
-
   return (
     <table className="bg-slate-50 col-span-4 mt-4 overflow-hidden rounded-lg shadow-md text-slate-800 w-full">
       <thead className="bg-slate-100 text-slate-900">
-        {table.getHeaderGroups().map((headerGroup) => (
-          <tr key={headerGroup.id}>
-            {headerGroup.headers.map((header) => (
-              <th key={header.id} className="px-4 py-2">
-                {!header.isPlaceholder &&
-                  flexRender(
-                    header.column.columnDef.header,
-                    header.getContext()
-                  )}
-              </th>
-            ))}
-          </tr>
-        ))}
+        <tr>
+          <th className="px-4 py-2">Name</th>
+          <th className="px-4 py-2">Start Date</th>
+          <th className="px-4 py-2">End Date</th>
+          <th className="px-4 py-2">Team</th>
+          <th className="px-4 py-2">Ticket Count</th>
+          <th />
+        </tr>
       </thead>
       <tbody>
-        {isLoading ? (
-          <tr className="border-b">
-            <td className="px-2 py-1">Loading...</td>
-          </tr>
-        ) : projects?.length === 0 ? (
+        {projects.length === 0 ? (
           <tr className="border-b">
             <td className="px-2 py-1">
               <span className="font-light italic text-slate-900">
@@ -132,13 +42,42 @@ export function ProjectList({
             </td>
           </tr>
         ) : (
-          table.getRowModel().rows.map((row) => (
-            <tr key={row.id} className="border-b">
-              {row.getVisibleCells().map((cell) => (
-                <td key={cell.id} className="px-2 py-1">
-                  {flexRender(cell.column.columnDef.cell, cell.getContext())}
-                </td>
-              ))}
+          projects.map((project) => (
+            <tr key={project.id} className="border-b">
+              <td className="px-2 py-1">
+                <div className="pl-1">{project.name}</div>
+              </td>
+              <td className="px-2 py-1">
+                <div className="text-center">
+                  {typeof project.startDate === 'undefined' ||
+                  project.startDate === null
+                    ? ''
+                    : project.startDate.toLocaleString('en-US', dateOptions)}
+                </div>
+              </td>
+              <td className="px-2 py-1">
+                <div className="text-center">
+                  {typeof project.endDate === 'undefined' ||
+                  project.endDate === null
+                    ? ''
+                    : project.endDate.toLocaleString('en-US', dateOptions)}
+                </div>
+              </td>
+              <td className="px-2 py-1">
+                <div className="team-avatars">
+                  {project.team.map(({ user }) => (
+                    <ProfileImage key={user.id} user={user.name ?? user.id} />
+                  ))}
+                </div>
+              </td>
+              <td className="px-2 py-1">
+                <div className="text-center">{project.tickets.length}</div>
+              </td>
+              <td className="px-2 py-1">
+                <Link href={`/projects/${project.id}`}>
+                  <InformationCircleIcon className="h-5 mx-auto w-5 hover:cursor-pointer" />
+                </Link>
+              </td>
             </tr>
           ))
         )}
