+++ apps/issues/pages/index_OLD.tsx
@@ -12,7 +12,7 @@ import {
 } from 'recharts';
 
 import Head from '../components/Head';
-import { ProjectList } from '../components/ProjectList';
+import { ProjectList } from '../components/ProjectList_OLD';
 import { trpc } from '../lib/trpc';
 import { requireLayoutProps } from '../lib/utils';
